+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Undocumented APIs should be considered internal and may change without warning.
 
 ## [10.0.0] 2023-02-24
 
+### Added
+
+-   `background-color` animations are now hardware accelerated.
+
 ### Removed
 
 -   Removing fallback for `IntersectionObserver`. Use a polyfill for support in older browsers.
@@ -16,6 +20,10 @@ Undocumented APIs should be considered internal and may change without warning.
 -   Using `exitBeforeEnter` with `AnimatePresence` now throws an error.
 -   Using `value.onChange` will now throw a warning with instructions to change to `value.on("change", callback)`.
 -   Using `AnimateSharedLayout` now throws an error.
+
+### Fixed
+
+-   `repeat: Infinity` no longer de-opts from pre-generated WAAPI animations.
 
 ## [9.1.7] 2023-02-24